apply changes made by Remi to move to new search api
